DRY up `configure` commands to use inheritance
These commands are essentially slightly different versions of doing the
same thing, so using inheritance fits well here.
