+++ src/commands/configure/group.rb
@@ -1,6 +1,5 @@
 
-require 'base_command'
-require 'configurator'
+require 'configure_command'
 require 'constants'
 
 
@@ -8,35 +7,20 @@ module Metalware
   module Commands
     module Configure
 
-      class Group < BaseCommand
+      class Group < ConfigureCommand
         def setup(args, _options)
           @group_name = args.first
         end
 
-        def run
-          configurator.configure
-        end
+        protected
 
-        def handle_interrupt(_e)
-          abort 'Exiting without saving...'
+        def answers_file
+          File.join(Constants::ANSWERS_PATH, 'groups', group_name)
         end
 
         private
 
         attr_reader :group_name
-
-        def configurator
-          Configurator.new(
-            highline: self,
-            configure_file: config.configure_file,
-            questions: :group,
-            answers_file: answers_file
-          )
-        end
-
-        def answers_file
-          File.join(Constants::ANSWERS_PATH, 'groups', group_name)
-        end
       end
 
     end
